IMPORTANT: I've changed the code from SourceHolderConstruction when copying over from Mendenhall's. In the construction of source foil coating, he aliases int as Side (seemingly object) and assigns EAST = 0, WEST = 1. I couldn't find these aliases and wasn't sure if that was exactly what he meant. I used int in the for loop, set the numbers as 0 and 1, and wrote my own string Append method.
As well, he uses ssign(sd) presumably for -1 or 1, to place two instances of coating symmetrically.
The method ssign is not declared (can't see how he got away with it) so I wrote two if statements to take
care of both cases. This looks to be about right, although still can't see in the geometry cause
the mother volume (container) hasn't been placed yet.